move open/close logic to `CreateContractFlow`
